% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{CIELAB_space}
\alias{CIELAB_space}
\title{Return CIELAB outline space coordinates and colours}
\usage{
CIELAB_space(RGB_points = RGB_space)
}
\arguments{
\item{RGB_points}{RGB data to move to UCIE. Default is set of points
on RGB extremes cube edges with step size 32.}
}
\value{
A data frame with columns L/A/B/colour, containing coordinates and hex codes.
}
\description{
Return CIELAB outline space coordinates and colours
}
\keyword{internal}
