% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{data2cielab}
\alias{data2cielab}
\title{Mapping 3D Data into CIELab color space}
\usage{
data2cielab(
  dataset,
  WL = 1,
  Wa = 1,
  Wb = 1,
  S = 1,
  LAB_coordinates = FALSE,
  center = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dataset}{3-column dataset to be translated into colors.}

\item{WL}{Weight of L* axis in optimization function. Default value 1.}

\item{Wa}{Weight of a* axis in optimization function. Default value 1.}

\item{Wb}{Weight of b* axis in optimization function. Default value 1.}

\item{S}{Scaling factor for color mapping. Default value 1.}

\item{LAB_coordinates}{Logical. If FALSE, the function returns a data frame with hex colors.
If TRUE, the function returns a data frame with the L*a*b* coordinates. Default value FALSE.}

\item{center}{Boolean: whether to center the data and color space during the
fitting process. Generally improves fit performance. Default is TRUE.}

\item{verbose}{Boolean: whether to print progress updates. Default is FALSE.}
}
\value{
Returns a data frame with 2 or 4 columns: a rownames column and either
hex colour codes or L/A/B coordinates.
}
\description{
Mapping 3D Data into CIELab color space
}
\examples{
\donttest{
  df <- data.frame(V1 = runif(10, 0, 1), V2 = runif(10, 0, 5), V3 = runif(10, 0, 30))
  data2cielab(df, Wb = 1.2, S = 1.6)
  data2cielab(df, LAB_coordinates = TRUE)
}
}
