% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{Distance}
\alias{Distance}
\title{Calculate distance of outside points to colour space.}
\usage{
Distance(S, RotL, Rota, Rotb, TrL, Tra, Trb, WL, Wa, Wb, Query, polygon, faces)
}
\arguments{
\item{S}{Scaling parameter.}

\item{RotL}{L* rotation parameter.}

\item{Rota}{a* rotation parameter.}

\item{Rotb}{b* rotation parameter.}

\item{TrL}{L* translation parameter.}

\item{Tra}{a* translation parameter.}

\item{Trb}{b* translation parameter.}

\item{WL}{Weight of L* axis.}

\item{Wa}{Weight of a* axis.}

\item{Wb}{Weight of b* axis.}

\item{Query}{(Convex hull of the) data to fit, as a matrix.}

\item{polygon}{Vertices (points) of the colour space convex hull, as a matrix.
Like output returned by [DataConvex()].}

\item{faces}{Edges (lines) of the colour space convex hull, as a matrix.
Like output returned by [geometry::convhulln()].}
}
\value{
A vector with distance to the colour space, for each point outside the space,
or 0 if no points are outside the space.
}
\description{
Determines which points are outside the colour space convex hull, calculates
the distance matrix between those points and the colour space, then, for each
data point, returns the distance to the closest colour space hull point.
}
\keyword{internal}
