% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{FitColorsFunction}
\alias{FitColorsFunction}
\title{Fit data to CIELAB space}
\usage{
FitColorsFunction(dataset, WL, Wa, Wb, center = TRUE)
}
\arguments{
\item{dataset}{Dataset to fit, as a matrix.}

\item{WL}{Weight of L* axis.}

\item{Wa}{Weight of a* axis.}

\item{Wb}{Weight of b* axis.}

\item{center}{Boolean: whether to center the data and color space during the
fitting process, which improves the fit performance but means the data needs
to be centered before the returned parameters can be used. Default is TRUE.}
}
\value{
A named vector of fitting parameters (S, RotL, Rota, Rotb, TrL, Tra, Trb).
}
\description{
Fit data to CIELAB space
}
\keyword{internal}
