% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{ObjectiveFunction}
\alias{ObjectiveFunction}
\title{Data fit minimalisation function}
\usage{
ObjectiveFunction(param, WL, Wa, Wb, data, polygon, faces)
}
\arguments{
\item{param}{Vector of fitting parameters (S, RotL, Rota, Rotb, TrL, Tra, Trb)}

\item{WL}{Weight of L* axis.}

\item{Wa}{Weight of a* axis.}

\item{Wb}{Weight of b* axis.}

\item{data}{(Convex hull of the) data to fit, as a matrix.}

\item{polygon}{Vertices (points) of the colour space convex hull, as a matrix.
Like output returned by [DataConvex()].}

\item{faces}{Edges (lines) of the colour space convex hull, as a matrix.
Like output returned by [geometry::convhulln()].}
}
\description{
Takes fit and weighting parameters, the convex hull of the data, and the
convex hull (vertices and faces) of the colour space to fit the data to.
Returns a fit metric, consisting of the scaling parameter, penalised strongly
by any points outside the colour space.
}
\details{
The penalty function calculates the distance of all outside points to the
closest colour space hull point, then subtracts the sum of all squared distances.
}
\keyword{internal}
