% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{PolygonCentroid}
\alias{PolygonCentroid}
\title{Calculate centroid of the data's convex hull polygon}
\usage{
PolygonCentroid(dataset)
}
\arguments{
\item{dataset}{A 3D matrix of either full data, or only the points making up
that data's convex hull (the vertices, like from [DataConvex()]).}
}
\description{
Calculates the centroid of a given data matrix's convex hull, using Delaunay triangulation.
}
\details{
Effectively finds the centroids of the tetrahedrons returned by Delaunay triangulation
and their volume. The weighted mean of those centroids, weighted by volume, is
the polygon's centroid. Unlike the arithmetic mean of the convex hull's vertices,
it is not biased towards more 'curved' sides that have more vertices.
}
\keyword{internal}
