% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{RGB2Lab}
\alias{RGB2Lab}
\title{Convert RGB coordinates to LAB coordinates}
\usage{
RGB2Lab(data)
}
\arguments{
\item{data}{tibble, data frame or matrix with columns `R`/`G`/`B`,
scaled 0-1 or 0-255}
}
\value{
A data frame or tibble (if supplied), otherwise a matrix, with columns
`L`/`A`/`B`.
}
\description{
Convert RGB coordinates to LAB coordinates
}
\keyword{internal}
