% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_transformations.R
\name{TranslationGuess}
\alias{TranslationGuess}
\title{Estimate an initial translation parameter}
\usage{
TranslationGuess(dataset, colorspace)
}
\arguments{
\item{dataset}{Data or convex hull vertices of data to fit inside the colorspace.}

\item{colorspace}{Colorspace points or convex hull vertices of colorspace.}
}
\value{
A named vector TrL,Tra,Trb of translation guess values.
}
\description{
Estimate an initial translation parameter
}
\details{
Centers using polygon centroid of any supplied data's convex hull,
so any version of the data (from full data to only points on convex hull) can be
used for both `dataset` and `colorspace`.
Should return basically 0,0,0 if both have been centered before guessing.
}
\keyword{internal}
