\name{ucie3DTransformation}
\alias{ucie3DTransformation}
\title{A trnaslation of 3D data into colors, using CIELab color space.}
\usage{
ucie3DTransformation(dataset, WL, Wa, Wb, S)
}
\description{
Translate the 3D data into colors of CIELab color space. CIELAB was intended as a perceptually uniform space, where a given numerical change corresponds to similar perceived change in color.
}
\examples{
ucie3DTransformation()
}
