\name{ucie3D}
\alias{hello}
\title{3D data as colors of CIELab color space.}
\usage{
ucie3D()
}
\description{
Translate the 3D data into colors of CIELab color space. CIELAB was intended as a perceptually uniform space, where a given numerical change corresponds to similar perceived change in color.
}
\examples{
ucie3D()
}
